pyproject.toml
[project]
name = "despesacerta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.122.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
]
